Start the schematics by connecting the HAT pins to the ESP32
The VCC on the e-Paper HAT is connected to 5V[1], but we will power it from
the 3V3 rails as I did in all my tests*. Because of that, we can't
replicate the pinout of the raspberry Pi header exactly. Instead, we
connect 3V3 to where 5V is on the Pi and leave the pins where the Pi has
3V3 unconnected (even though they're not routed in the HAT).

* The level shifter in the HAT (a TXS0108E[2]) takes from 1.65 to 5.5V on the
B port, where VCC is routed.

[1]: https://www.waveshare.com/w/upload/8/87/E-Paper-Driver-HAT-Schematic.pdf
[2]: https://www.ti.com/lit/ds/symlink/txs0108e.pdf?ts=1611473568779